filter the results according to the query
